+++ config/config/src/Config.ts
@@ -157,6 +157,7 @@ export interface Config {
   gitShallowHosts?: string[]
   legacyDirFiltering?: boolean
   onlyBuiltDependencies?: string[]
+  dedupePeerDependents?: boolean
 
   registries: Registries
   ignoreWorkspaceRootCheck: boolean