+++ config/plugin-commands-config/src/ConfigCommandOptions.ts
@@ -2,6 +2,7 @@ import { Config } from '@pnpm/config'
 
 export type ConfigCommandOptions = Pick<Config,
 | 'configDir'
+| 'cliOptions'
 | 'dir'
 | 'global'
 | 'rawConfig'