+++ pkg-manager/core/src/install/index.ts
@@ -45,6 +45,7 @@ import {
   DependenciesGraphNode,
   PinnedVersion,
   resolveDependencies,
+  UpdateMatchingFunction,
   WantedDependency,
 } from '@pnpm/resolve-dependencies'
 import {
@@ -89,36 +90,50 @@ const BROKEN_LOCKFILE_INTEGRITY_ERRORS = new Set([
 
 const DEV_PREINSTALL = 'pnpm:devPreinstall'
 
-export type DependenciesMutation = (
-  {
-    mutation: 'install'
-    pruneDirectDependencies?: boolean
-  } | {
-    allowNew?: boolean
-    dependencySelectors: string[]
-    mutation: 'installSome'
-    peer?: boolean
-    pruneDirectDependencies?: boolean
-    pinnedVersion?: PinnedVersion
-    targetDependenciesField?: DependenciesField
-  } | {
-    mutation: 'uninstallSome'
-    dependencyNames: string[]
-    targetDependenciesField?: DependenciesField
-  } | {
-    mutation: 'unlink'
-  } | {
-    mutation: 'unlinkSome'
-    dependencyNames: string[]
-  }
-)
+interface InstallMutationOptions {
+  update?: boolean
+  updateMatching?: UpdateMatchingFunction
+  updatePackageManifest?: boolean
+}
+
+export interface InstallDepsMutation extends InstallMutationOptions {
+  mutation: 'install'
+  pruneDirectDependencies?: boolean
+}
+
+export interface InstallSomeDepsMutation extends InstallMutationOptions {
+  allowNew?: boolean
+  dependencySelectors: string[]
+  mutation: 'installSome'
+  peer?: boolean
+  pruneDirectDependencies?: boolean
+  pinnedVersion?: PinnedVersion
+  targetDependenciesField?: DependenciesField
+}
+
+export interface UninstallSomeDepsMutation {
+  mutation: 'uninstallSome'
+  dependencyNames: string[]
+  targetDependenciesField?: DependenciesField
+}
+
+export interface UnlinkDepsMutation {
+  mutation: 'unlink'
+}
+
+export interface UnlinkSomeDepsMutation {
+  mutation: 'unlinkSome'
+  dependencyNames: string[]
+}
+
+export type DependenciesMutation = InstallDepsMutation | InstallSomeDepsMutation | UninstallSomeDepsMutation | UnlinkDepsMutation | UnlinkSomeDepsMutation
 
 export async function install (
   manifest: ProjectManifest,
   opts: Omit<InstallOptions, 'allProjects'> & {
     preferredVersions?: PreferredVersions
     pruneDirectDependencies?: boolean
-  }
+  } & InstallMutationOptions
 ) {
   const rootDir = opts.dir ?? process.cwd()
   const projects = await mutateModules(
@@ -127,6 +142,9 @@ export async function install (
         mutation: 'install',
         pruneDirectDependencies: opts.pruneDirectDependencies,
         rootDir,
+        update: opts.update,
+        updateMatching: opts.updateMatching,
+        updatePackageManifest: opts.updatePackageManifest,
       },
     ],
     {
@@ -165,10 +183,17 @@ export async function mutateModulesInSingleProject (
     rootDir: string
     modulesDir?: string
   },
-  maybeOpts: Omit<MutateModulesOptions, 'allProjects'>
+  maybeOpts: Omit<MutateModulesOptions, 'allProjects'> & InstallMutationOptions
 ): Promise<UpdatedProject> {
   const [updatedProject] = await mutateModules(
-    [project],
+    [
+      {
+        ...project,
+        update: maybeOpts.update,
+        updateMatching: maybeOpts.updateMatching,
+        updatePackageManifest: maybeOpts.updatePackageManifest,
+      } as MutatedProject,
+    ],
     {
       ...maybeOpts,
       allProjects: [{
@@ -195,7 +220,7 @@ export async function mutateModules (
     throw new PnpmError('OPTIONAL_DEPS_REQUIRE_PROD_DEPS', 'Optional dependencies cannot be installed without production dependencies')
   }
 
-  const installsOnly = projects.every((project) => project.mutation === 'install')
+  const installsOnly = projects.every((project) => project.mutation === 'install' && !project.update && !project.updateMatching)
   if (!installsOnly) opts.strictPeerDependencies = false
   // @ts-expect-error
   opts['forceNewModules'] = installsOnly
@@ -306,7 +331,6 @@ export async function mutateModules (
       opts.frozenLockfileIfExists && ctx.existsWantedLockfile
     if (
       !ctx.lockfileHadConflicts &&
-      !opts.update &&
       !opts.fixLockfile &&
       !opts.dedupe &&
       installsOnly &&
@@ -396,7 +420,9 @@ export async function mutateModules (
           if (BROKEN_LOCKFILE_INTEGRITY_ERRORS.has(error.code)) {
             needsFullResolution = true
             // Ideally, we would not update but currently there is no other way to redownload the integrity of the package
-            opts.update = true
+            for (const project of projects) {
+              (project as InstallMutationOptions).update = true
+            }
           }
           // A broken lockfile may be caused by a badly resolved Git conflict
           logger.warn({
@@ -432,14 +458,14 @@ export async function mutateModules (
       case 'install': {
         await installCase({
           ...projectOpts,
-          updatePackageManifest: opts.updatePackageManifest ?? opts.update,
+          updatePackageManifest: (projectOpts as InstallDepsMutation).updatePackageManifest ?? (projectOpts as InstallDepsMutation).update,
         })
         break
       }
       case 'installSome': {
         await installSome({
           ...projectOpts,
-          updatePackageManifest: opts.updatePackageManifest !== false,
+          updatePackageManifest: (projectOpts as InstallSomeDepsMutation).updatePackageManifest !== false,
         })
         break
       }
@@ -510,7 +536,7 @@ export async function mutateModules (
       const wantedDependencies = getWantedDependencies(project.manifest, {
         autoInstallPeers: opts.autoInstallPeers,
         includeDirect: opts.includeDirect,
-        updateWorkspaceDependencies: opts.update,
+        updateWorkspaceDependencies: project.update,
         nodeExecPath: opts.nodeExecPath,
       })
         .map((wantedDependency) => ({ ...wantedDependency, updateSpec: true, preserveNonSemverVersionSpec: true }))
@@ -549,7 +575,7 @@ export async function mutateModules (
         devDependencies,
         optional: project.targetDependenciesField === 'optionalDependencies',
         optionalDependencies,
-        updateWorkspaceDependencies: opts.update,
+        updateWorkspaceDependencies: project.update,
         preferredSpecs,
         overrides: opts.overrides,
       })
@@ -684,7 +710,7 @@ export async function addDependenciesToPackage (
     peer?: boolean
     pinnedVersion?: 'major' | 'minor' | 'patch'
     targetDependenciesField?: DependenciesField
-  }
+  } & InstallMutationOptions
 ) {
   const rootDir = opts.dir ?? process.cwd()
   const projects = await mutateModules(
@@ -697,6 +723,9 @@ export async function addDependenciesToPackage (
         pinnedVersion: opts.pinnedVersion,
         rootDir,
         targetDependenciesField: opts.targetDependenciesField,
+        update: opts.update,
+        updateMatching: opts.updateMatching,
+        updatePackageManifest: opts.updatePackageManifest,
       },
     ],
     {
@@ -798,8 +827,9 @@ const _installInContext: InstallFunction = async (projects, ctx, opts) => {
     stage: 'resolution_started',
   })
 
+  const update = projects.some((project) => (project as InstallMutationOptions).update)
   const preferredVersions = opts.preferredVersions ?? (
-    !opts.update
+    !update
       ? getPreferredVersionsFromLockfileAndManifests(ctx.wantedLockfile.packages, Object.values(ctx.projects).map(({ manifest }) => manifest))
       : undefined
   )
@@ -807,7 +837,8 @@ const _installInContext: InstallFunction = async (projects, ctx, opts) => {
     !opts.currentLockfileIsUpToDate ||
     opts.force ||
     opts.needsFullResolution ||
-    ctx.lockfileHadConflicts
+    ctx.lockfileHadConflicts ||
+    opts.dedupePeerDependents
 
   // Ignore some fields when fixing lockfile, so these fields can be regenerated
   // and make sure it's up to date
@@ -852,7 +883,8 @@ const _installInContext: InstallFunction = async (projects, ctx, opts) => {
       allowNonAppliedPatches: opts.allowNonAppliedPatches,
       autoInstallPeers: opts.autoInstallPeers,
       currentLockfile: ctx.currentLockfile,
-      defaultUpdateDepth: (opts.update || (opts.updateMatching != null)) ? opts.depth : -1,
+      defaultUpdateDepth: opts.depth,
+      dedupePeerDependents: opts.dedupePeerDependents,
       dryRun: opts.lockfileOnly,
       engineStrict: opts.engineStrict,
       force: opts.force,
@@ -873,7 +905,6 @@ const _installInContext: InstallFunction = async (projects, ctx, opts) => {
       saveWorkspaceProtocol: opts.saveWorkspaceProtocol,
       storeController: opts.storeController,
       tag: opts.tag,
-      updateMatching: opts.updateMatching,
       virtualStoreDir: ctx.virtualStoreDir,
       wantedLockfile: ctx.wantedLockfile,
       workspacePackages: opts.workspacePackages,
@@ -1246,7 +1277,9 @@ const installInContext: InstallFunction = async (projects, ctx, opts) => {
     ) throw error
     opts.needsFullResolution = true
     // Ideally, we would not update but currently there is no other way to redownload the integrity of the package
-    opts.update = true
+    for (const project of projects) {
+      (project as InstallMutationOptions).update = true
+    }
     logger.warn({
       error,
       message: error.message,