+++ cli/default-reporter/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @pnpm/default-reporter
 
+## 11.0.38
+
+### Patch Changes
+
+- @pnpm/config@16.7.1
+
+## 11.0.37
+
+### Patch Changes
+
+- Updated dependencies [5c31fa8be]
+  - @pnpm/config@16.7.0
+
+## 11.0.36
+
+### Patch Changes
+
+- @pnpm/config@16.6.4
+
 ## 11.0.35
 
 ### Patch Changes