+++ hooks/read-package-hook/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pnpm/hooks.read-package-hook
 
+## 2.1.1
+
+### Patch Changes
+
+- d583fbb2a: Fixed issue where allowedVersions wouldn't apply correctly if only parent>child selectors were used
+
+## 2.1.0
+
+### Minor Changes
+
+- f39d608ac: Extends the `pnpm.peerDependencyRules.allowedVersions` `package.json` option to support the `parent>child` selector syntax. This syntax allows for extending specific `peerDependencies` [#6108](https://github.com/pnpm/pnpm/pull/6108).
+
 ## 2.0.12
 
 ### Patch Changes