+++ packages/plugin-commands-setup/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @pnpm/plugin-commands-setup
 
+## 3.0.38
+
+### Patch Changes
+
+- @pnpm/cli-utils@1.1.3
+
+## 3.0.37
+
+### Patch Changes
+
+- d80661d42: The configuration added by `pnpm setup` should check if the pnpm home directory is already in the PATH before adding to the PATH.
+
+  Before this change, this code was added to the shell:
+
+  ```sh
+  export PNPM_HOME="$HOME/Library/pnpm"
+  export PATH="$PNPM_HOME:$PATH"
+  ```
+
+  Now this will be added:
+
+  ```sh
+  export PNPM_HOME="$HOME/Library/pnpm"
+  case ":$PATH:" in
+    *":$PNPM_HOME:"*) ;;
+    *) export PATH="$PNPM_HOME:$PATH" ;;
+  esac
+  ```
+
+- Updated dependencies [7d64d757b]
+  - @pnpm/cli-utils@1.1.2
+
+## 3.0.36
+
+### Patch Changes
+
+- @pnpm/cli-utils@1.1.1
+
 ## 3.0.35
 
 ### Patch Changes