+++ releasing/plugin-commands-deploy/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @pnpm/plugin-commands-deploy
 
+## 2.0.38
+
+### Patch Changes
+
+- Updated dependencies [690bead26]
+  - @pnpm/plugin-commands-installation@11.5.3
+  - @pnpm/cli-utils@1.1.3
+
+## 2.0.37
+
+### Patch Changes
+
+- Updated dependencies [7d64d757b]
+  - @pnpm/cli-utils@1.1.2
+  - @pnpm/plugin-commands-installation@11.5.2
+
+## 2.0.36
+
+### Patch Changes
+
+- @pnpm/plugin-commands-installation@11.5.1
+- @pnpm/cli-utils@1.1.1
+
 ## 2.0.35
 
 ### Patch Changes