+++ reviewing/plugin-commands-outdated/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @pnpm/plugin-commands-outdated
 
+## 8.0.33
+
+### Patch Changes
+
+- @pnpm/config@16.7.1
+- @pnpm/cli-utils@1.1.3
+
+## 8.0.32
+
+### Patch Changes
+
+- Updated dependencies [7d64d757b]
+- Updated dependencies [5c31fa8be]
+  - @pnpm/cli-utils@1.1.2
+  - @pnpm/config@16.7.0
+
+## 8.0.31
+
+### Patch Changes
+
+- Updated dependencies [ed946c73e]
+  - @pnpm/lockfile-file@7.0.5
+  - @pnpm/outdated@11.0.20
+  - @pnpm/config@16.6.4
+  - @pnpm/cli-utils@1.1.1
+
 ## 8.0.30
 
 ### Patch Changes