+++ store/plugin-commands-server/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @pnpm/plugin-commands-server
 
+## 5.0.38
+
+### Patch Changes
+
+- @pnpm/config@16.7.1
+- @pnpm/cli-utils@1.1.3
+- @pnpm/store-connection-manager@5.2.16
+
+## 5.0.37
+
+### Patch Changes
+
+- Updated dependencies [7d64d757b]
+- Updated dependencies [5c31fa8be]
+  - @pnpm/cli-utils@1.1.2
+  - @pnpm/config@16.7.0
+  - @pnpm/store-connection-manager@5.2.15
+
+## 5.0.36
+
+### Patch Changes
+
+- @pnpm/config@16.6.4
+- @pnpm/cli-utils@1.1.1
+- @pnpm/store-connection-manager@5.2.14
+
 ## 5.0.35
 
 ### Patch Changes