+++ store/store-connection-manager/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @pnpm/store-connection-manager
 
+## 5.2.16
+
+### Patch Changes
+
+- @pnpm/config@16.7.1
+
+## 5.2.15
+
+### Patch Changes
+
+- Updated dependencies [5c31fa8be]
+  - @pnpm/config@16.7.0
+
+## 5.2.14
+
+### Patch Changes
+
+- @pnpm/config@16.6.4
+
 ## 5.2.13
 
 ### Patch Changes