+++ workspace/filter-workspace-packages/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @pnpm/filter-workspace-packages
 
+## 6.0.38
+
+### Patch Changes
+
+- @pnpm/find-workspace-packages@5.0.38
+
+## 6.0.37
+
+### Patch Changes
+
+- @pnpm/find-workspace-packages@5.0.37
+
+## 6.0.36
+
+### Patch Changes
+
+- @pnpm/find-workspace-packages@5.0.36
+
 ## 6.0.35
 
 ### Patch Changes